rename var-picker due to style overlap with author
