+++ elements/tests/tests.js
@@ -7,4 +7,4 @@ import '~/elements/a2j-template-ssr/a2j-template-ssr-test'
 import '~/elements/a2j-variable/a2j-variable-test'
 import '~/elements/options-pane/options-pane-test'
 import '~/elements/element-container/element-container-test'
-import '~/elements/varpicker/varpicker-test'
+import '~/elements/variable-picker/variable-picker-test'
