+++ elements/variable-picker/demo.html
@@ -1,18 +1,18 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>&lt;var-picker&gt;</title>
+    <title>&lt;variable-picker&gt;</title>
   </head>
 
   <body>
     <script id="template" type="text/stache" can-autorender>
       <can-import from="~/styles/" />
-      <can-import from="~/elements/varpicker/demo/" />
+      <can-import from="~/elements/variable-picker/demo/" />
 
       <div class="bootstrap-styles" style="margin: 20px;">Type 'c' to start search
         <div class="row clearfix" style="margin-top: 20px;">
           <div class="col-md-6 col-md-offset-3">
-            <varpicker-demo />
+            <variable-picker-demo />
           </div>
         </div>
       </div>
