+++ elements/variable-picker/demo/demo.stache
@@ -1,0 +1,3 @@
+<can-import from="~/elements/variable-picker/" />
+
+<variable-picker style="width: 200px;" vm:variables:from="variables" />
