+++ elements/variable-picker/variable-picker-test.html
@@ -8,7 +8,7 @@
     <div id="test-area"></div>
     <script src="../../node_modules/steal/steal.js"
       data-mocha="bdd"
-      data-main="~/elements/varpicker/varpicker-test">
+      data-main="~/elements/variable-picker/variable-picker-test">
     </script>
   </body>
 </html>
