+++ elements/variable-picker/variable-picker.less
@@ -2,8 +2,8 @@
 
 @var-highlight:   #def;
 
-var-picker {
-  .var-picker-warning {
+variable-picker {
+  .variable-picker-warning {
     color: #777777;
     background-color: #fdfaec;
     border-color: #e0e0e0;
