--- elements/varpicker/demo/demo.stache
@@ -1,3 +1,0 @@
-<can-import from="~/elements/varpicker/" />
-
-<var-picker style="width: 200px;" vm:variables:from="variables" />
